+++ ext/drawText/drawText Tests/ExtensionsTests.swift
@@ -1,5 +1,0 @@
-import XCTest
-
-class ExtensionsTests: DrawTextTest {
-
-   }